Move trigger injection onto ScheduledRtl
